Translate and add a denomination description to Field Order
To give users a better indication of where in a list of displayed fields
this will show up on the front-end of the site (end-user facing).
While this is a control targetting user facing fields only, it does not
limit by only those that are displayed, as it is plausible for a content
editor to wish to order fields in both the summary and detail views,
which may be different subsets of the available fields.