FIX when no data is set, show the ResourceLocatorField by default
The 'nomrmally hidden' status of the ResourceLocatorField is good,
but only when a resource has already been chosen and its columns
loaded. In the case that one has not been saved, the "Data" tab on
a Page loads empty. So in order to allow people to enter data and
save the columns initially, we need to show the field.