FIX show ResourceLocatorField description only if no error
Previously when showing an error message for one of the two fields
that make up the ResourceLocatorField component they would push down
the description, leaving a large empty space in the case that the
error was for the select field (as opposed to the left positioned
text input). To remedy this we can simply hide the description when
an error/validation message is showing for one of the inner fields,
effectively looking like the description was replaced.

To achieve this the fieldHolder HOC had to be removed to allow us
access to control whether or not the description is shown - this is
because the description is passed in via PHP, so cannot be re-written
on the fly. The communication to show or hide is then either directly
controlled by the render function, or passed 'up' from the select field
via component state.