FIX minor display bugs when site setup differs from expectations
Upon loading a page after having just created it, I was greeted by the
title "null / null" - and no content. This was upsetting because I had
entered a resource endpoint to load. The fix for the title was easy -
just don't return one if one of the requsites isn't set. The reason for
the lack of data loading was due to loading the development environment
from a sub-folder - in which we should correctly form relative URLs as
the base for react router. An example invalid route was removed.

Features added:
 Download link for fetching resource
 Link to resource page on CKAN host