ADD sort functionality when clicking column header
Griddle by default allows a sort of the columns by clicking the title of
a the column a user wishes to sort. Previously this did nothing, and
required us to connect the CKAN API to fetch a new page of data, having
the results sorted by the desired column first.